Screwing around trying to eke out a little more performance
Test 64 bits worth of cells (i.e. 8 cells) at a time in the
step function and only bother to look at the contents of the
individual cells if the whole doubleword is non-zero.

Amazingly, it appears to have reduced the running time of the
step() function by 40 % - from approx 50μs to approx 50μs - on
my Mac. If only there was a uint128_t type that the compiler
could compile into a LDP instruction to somehow magically make
the performance of skipping blank cells even better...