Try rendering to a pixel array and then rendering _that_ to the screen
Advantage is that you only need to update pixels that _change_ in a
particular generation. Disadvantage is that it takes around 3-6ms to
copy that array into a texture and render that.

Presumably most of the time is taken sending the pixels down to the
GPU. Presumably there isn't any way of getting that much data to the
GPU every frame without it taking a few ms. Or there is a way, but I
have no idea how modern graphics hardware works and consequently
haven't found it.